refactor: split up refund page
