+++ src/style/index.scss
@@ -651,3 +651,7 @@ textarea {
     font-weight: 600;
     text-align: center;
 }
+
+header {
+    margin-bottom: 1rem;
+}
