+++ tests/pages/Refund.spec.tsx
@@ -54,7 +54,7 @@ describe("Refund", () => {
         const refundFrame = (await screen.findByTestId(
             "refundFrame",
         )) as HTMLDivElement;
-        expect(refundFrame.children.length).toEqual(5);
+        expect(refundFrame.children.length).toEqual(7);
 
         const uploadInput = await screen.findByTestId("refundUpload");
         const swapFile = new File(["{}"], "swap.json", {
